Ratcheting up the warning level, addressing.  Moving many functions into headers and using auto.

+++ src/cc_omniview_canvas.h
@@ -76,7 +76,7 @@ public:
     void OnCmd_ToggleShowOnlySelected();
 
 private:
-    MarcherInfo GetMarcherInfo(size_t which) const;
+    MarcherInfo GetMarcherInfo(int which) const;
     std::multimap<double, MarcherInfo> ParseAndDraw3dMarchers() const;
 
     std::shared_ptr<CCOmniView_GLContext> m_glContext;
